Fix deadlock in EventTypeChannel
Experiment was sometimes freezing due to deadlock in EventTypeChannel:

- Most of the threads are at EventTypeChannel.java:155 waiting for the lock on
  equalTypesEventProcessorMap.

- One thread has the lock on equalTypesEventProcessorMap and is waiting for the
  lock on commands at AbstractExperiment.java:191.

- Another thread has the lock on commands and is waiting for the lock on
  equalTypesEventProcessorMap at EventTypeChannel.java:155.

- The latter two threads are deadlocked, because each needs to aquire the lock
  held by the other before releasing the lock it currently holds.

The solution I implemented was switching to ReadWriteLocks for locking
equalTypesEventProcessorMap and the similar acceptsSubtypesEventProcessors,
instead of synchronized blocks. This allows any number of concurrent reads,
while giving exclusive access to the thread with the write lock. (Reads are all
that are happening with the deadlocking code).